refactor(tests): resolve an attribute's receiver, and split the helper (#324)
Three changes, and the first is a defect the subprocess-lifecycle lane
found in the helper itself.

RESOLVING A RECEIVER. After `class G: mod = os`, the call `G.mod.getpgid`
has no known prefix and its outermost attribute is `getpgid`, which
nothing bound, so `Bindings.resolve` answered None. Under the earlier
three-branch `_classify_call` that was a SILENT miss, which is exactly the
defect #324 exists to end, reproduced one level up. The no-third-case rule
had already turned it into a reported undecided row; `_through_attribute`
now resolves the receiver first and answers it outright. Its bound is
pinned beside it: `other.lookup`, whose receiver nothing bound, stays in
the reported half rather than becoming a false `seen`. Both directions
have a test, and deleting the recursive step turns the first one red.

Measured over kstrl/: the subprocess spawn inventory is unchanged at 68
seen and 12 undecided, and the pgid inventory is unchanged at one seen.
So the step adds reach without moving a package row.

`Bindings.attributes` was never meant to resolve receivers. It maps a
bare attribute NAME to what some receiver bound it to, a deliberate
over-match for two rows the AST cannot type. Receiver resolution is a
second mechanism, and this is it.

THE SPLIT. The file-length ratchet is right that 800 lines is more than
one job, and it had no headroom left for the two sweeps that will adopt
it. `astwalk` is now a package: `corpus` (the files and the folding),
`net` (census and the total partition, neither enumerating a node type),
`resolve` (the layer that can go blind), `scope`, `disclose`. Every name
is re-exported from `__init__`, so `from tests.helpers import astwalk`
and `from tests.helpers.astwalk import X` both still resolve and no
caller changes. The move broke `REPO_ROOT`, and the corpus test caught it
rather than the suite silently going green over an empty file list.

THE TIMEOUT GUARD. `tests/test_timeout_enforcement.py` drops
`_subprocess_aliases` and `_popen_names` for `astwalk.resolved_calls`.
Spawn sites 68 before, 68 after, none lost and none gained; `popen_bound`
in `agents/proc.py` gains `self._proc`, a widening. Three shapes leave
its miss ledger for a new catch table (`_P = subprocess.Popen`, the
annotated form, and `getattr(subprocess, "Popen")`), the remaining two
are genuine scope rows, and the 12 undecided rows are now pinned.
